Apply 2to3 conversions and fixup failing tests
Tests failed with test_client returning bytes instead of string. Doing response.get_data(as_text=True)
returns a string object as expected by the tests. It also seems like one of the tests were expecting
the query parameters to be in a determined order. Not sure if this is necessary on Atlassian's side
but I believe it is the HTTP standard that the ordering should not matter.